Editing home screen and it's content
Changing NestedScrollView to CustomScrollView this will improve the performance. Because  NestedScrollView  is manages two separate scrollable areas
we don't need this in our case + CustomScrollView has SliverAppBar so we can use it for the fixed part in our code.

Home Content we removed the CustomScrollView as it should be in the Home Screen not here and insted of it we used SliverGrid, then called the Widget inside slivers [ the parent Widget ]