adding drawer and adddress
First we have removed the setting icon from the layout, we have created a drawer instead. Created a ModalBottomSheet for the address when placing an order, now it's working fine but local. Created an address model.

Next will work on setting and profile when we are done will work on the payment.